+++ changelog.md
@@ -15,6 +15,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityTablePanel, EDIT control used as double click action if available.
 - TabbedDetailLayout bug fixed, detail panel no longer loses focus when embedded, focus flicker reduced by clearing the focus.
+- EntityDialogs.DefaultEditEntityDialogBuilder bug fixed, now reverts any edits on cancel.
 
 ## 0.18.7
 ### is.codion.common.core